🔖 Release: v3.6.0 버전 릴리즈 (#463)
- Feat: 관리자 전용 API 관련 다수의 추가 및 변경 등
- Refactor: @securityrequirement를 MemberAdminRestController의 클래스 수준에서 선언
- Refactor: 댓글 REST 컨트롤러에서 @operation 어노테이션에 security 속성 추가

+++ src/main/java/kr/modusplant/domains/member/adapter/controller/MemberAdminController.java
@@ -2,7 +2,9 @@ package kr.modusplant.domains.member.adapter.controller;
 
 import kr.modusplant.domains.member.adapter.helper.MemberValidationHelper;
 import kr.modusplant.domains.member.domain.event.CommentAbuseReportApproveEvent;
+import kr.modusplant.domains.member.domain.event.CommentAbuseReportDismissEvent;
 import kr.modusplant.domains.member.domain.event.PostAbuseReportApproveEvent;
+import kr.modusplant.domains.member.domain.event.PostAbuseReportDismissEvent;
 import kr.modusplant.domains.member.domain.vo.*;
 import kr.modusplant.domains.member.usecase.model.read.CommentAbuseReportDashboardReadModel;
 import kr.modusplant.domains.member.usecase.model.read.PostAbuseReportDashboardReadModel;
@@ -10,6 +12,9 @@ import kr.modusplant.domains.member.usecase.model.read.ProposalOrBugReportDashbo
 import kr.modusplant.domains.member.usecase.port.repository.ReportDashboardRepository;
 import kr.modusplant.domains.member.usecase.port.repository.ReportRepository;
 import kr.modusplant.domains.member.usecase.record.*;
+import kr.modusplant.domains.member.usecase.response.CommentAbuseReportDashboardResponse;
+import kr.modusplant.domains.member.usecase.response.PostAbuseReportDashboardResponse;
+import kr.modusplant.domains.member.usecase.response.ProposalOrBugReportDashboardResponse;
 import kr.modusplant.shared.exception.ExistsValueException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -32,13 +37,32 @@ public class MemberAdminController {
     private final ReportDashboardRepository reportDashboardRepository;
     private final ApplicationEventPublisher applicationEventPublisher;
 
-    public List<ProposalOrBugReportDashboardReadModel> getProposalOrBug(ProposalOrBugReportGetRecord record) {
-        if (record.lastReportUlid() != null) {
+    public ProposalOrBugReportDashboardResponse getProposalOrBug(ProposalOrBugReportGetRecord record) {
+        int reportPageSize = record.size() + 1;
+        List<ProposalOrBugReportDashboardReadModel> readModels;
+        if (record.lastReportUlid() != null) { // 특정 lastReportId 이후부터 조회
             ReportId reportId = ReportId.create(record.lastReportUlid());
             memberValidationHelper.validateIfReportExists(reportId);
-            return reportDashboardRepository.getProposalOrBugReports(ReportPageSize.create(record.size()), record.status(), reportId);
-        } else {
-            return reportDashboardRepository.getProposalOrBugReports(ReportPageSize.create(record.size()), record.status(), null);
+            readModels = reportDashboardRepository.getProposalOrBugReports(
+                    ReportPageSize.create(reportPageSize), record.status(), reportId);
+        } else { // 첫 보고서 데이터부터 조회
+            readModels = reportDashboardRepository.getProposalOrBugReports(
+                    ReportPageSize.create(reportPageSize), record.status(), null);
+        }
+
+        if (readModels.size() == reportPageSize) { // hasNext == true
+            List<ProposalOrBugReportDashboardReadModel> returnedReadModels =
+                    readModels.subList(0, reportPageSize - 1);
+            return ProposalOrBugReportDashboardResponse.of(
+                    returnedReadModels, returnedReadModels.getLast().ulid(), true);
+        } else { // hasNext == false
+            if (!readModels.isEmpty()) {
+                return ProposalOrBugReportDashboardResponse.of(
+                        readModels, readModels.getLast().ulid(), false);
+            } else {
+                return ProposalOrBugReportDashboardResponse.of(
+                        readModels, null, false);
+            }
         }
     }
 
@@ -54,19 +78,39 @@ public class MemberAdminController {
     public void removeProposalOrBug(ProposalOrBugReportRemoveRecord record) {
         ReportId reportId = ReportId.create(record.reportUlid());
         if (reportRepository.isIdExistInProposalOrBugReport(reportId)) {
+            if (reportRepository.isUncheckedInProposalOrBugReport(reportId)) {
+                throw new ExistsValueException(NOT_FOUND_PROPOSAL_OR_BUG_REPORT_CHECKED, "checkedAt");
+            }
             reportRepository.removeProposalOrBugReport(reportId);
         }
     }
 
-    public List<PostAbuseReportDashboardReadModel> getPostAbuseReport(PostAbuseReportGetRecord record) {
+    public PostAbuseReportDashboardResponse getPostAbuseReport(PostAbuseReportGetRecord record) {
+        int reportPageSize = record.size() + 1;
+        List<PostAbuseReportDashboardReadModel> readModels;
         if (record.lastPostUlid() != null) {
             ActivitySubjectPostId activitySubjectPostId = ActivitySubjectPostId.create(record.lastPostUlid());
             memberValidationHelper.validateIfActivitySubjectPostExists(activitySubjectPostId);
-            return reportDashboardRepository.getPostAbuseReports(
-                    ReportPageSize.create(record.size()), record.status(), activitySubjectPostId);
+            readModels = reportDashboardRepository.getPostAbuseReports(
+                    ReportPageSize.create(reportPageSize), record.status(), activitySubjectPostId);
         } else {
-            return reportDashboardRepository.getPostAbuseReports(
-                    ReportPageSize.create(record.size()), record.status(), null);
+            readModels = reportDashboardRepository.getPostAbuseReports(
+                    ReportPageSize.create(reportPageSize), record.status(), null);
+        }
+
+        if (readModels.size() == reportPageSize) { // hasNext == true
+            List<PostAbuseReportDashboardReadModel> returnedReadModels =
+                    readModels.subList(0, reportPageSize - 1);
+            return PostAbuseReportDashboardResponse.of(
+                    returnedReadModels, returnedReadModels.getLast().ulid(), true);
+        } else { // hasNext == false
+            if (!readModels.isEmpty()) {
+                return PostAbuseReportDashboardResponse.of(
+                        readModels, readModels.getLast().ulid(), false);
+            } else {
+                return PostAbuseReportDashboardResponse.of(
+                        readModels, null, false);
+            }
         }
     }
 
@@ -76,6 +120,7 @@ public class MemberAdminController {
         if (reportDashboardRepository.isDismissedInPostAbuseReportDashboard(postId)) {
             throw new ExistsValueException(EXISTS_POST_ABUSE_REPORT_DISMISSED, "status");
         }
+        applicationEventPublisher.publishEvent(PostAbuseReportDismissEvent.create(postId.getValue()));
         return reportDashboardRepository.dismissPostAbuseReport(postId);
     }
 
@@ -89,18 +134,38 @@ public class MemberAdminController {
         return reportDashboardRepository.approvePostAbuseReport(postId);
     }
 
-    public List<CommentAbuseReportDashboardReadModel> getCommentAbuseReport(CommentAbuseReportGetRecord record) {
+    public CommentAbuseReportDashboardResponse getCommentAbuseReport(CommentAbuseReportGetRecord record) {
+        int reportPageSize = record.size() + 1;
+        List<CommentAbuseReportDashboardReadModel> readModels;
         if (!(record.lastPostUlid() == null && record.lastPath() == null)) {
             ActivitySubjectCommentId activitySubjectCommentId =
                     ActivitySubjectCommentId.create(
                             ActivitySubjectPostId.create(record.lastPostUlid()),
                             ActivitySubjectCommentPath.create(record.lastPath()));
             memberValidationHelper.validateIfActivitySubjectCommentExists(activitySubjectCommentId);
-            return reportDashboardRepository.getCommentAbuseReports(
-                    ReportPageSize.create(record.size()), record.status(), activitySubjectCommentId);
+            readModels = reportDashboardRepository.getCommentAbuseReports(
+                    ReportPageSize.create(reportPageSize), record.status(), activitySubjectCommentId);
         } else {
-            return reportDashboardRepository.getCommentAbuseReports(
-                    ReportPageSize.create(record.size()), record.status(), null);
+            readModels = reportDashboardRepository.getCommentAbuseReports(
+                    ReportPageSize.create(reportPageSize), record.status(), null);
+        }
+
+        if (readModels.size() == reportPageSize) { // hasNext == true
+            List<CommentAbuseReportDashboardReadModel> returnedReadModels =
+                    readModels.subList(0, reportPageSize - 1);
+            return CommentAbuseReportDashboardResponse.of(
+                    returnedReadModels,
+                    returnedReadModels.getLast().postUlid(),
+                    returnedReadModels.getLast().path(),
+                    true);
+        } else { // hasNext == false
+            if (!readModels.isEmpty()) {
+                return CommentAbuseReportDashboardResponse.of(
+                        readModels, readModels.getLast().postUlid(), readModels.getLast().path(), false);
+            } else {
+                return CommentAbuseReportDashboardResponse.of(
+                        readModels, null, null, false);
+            }
         }
     }
 
@@ -112,6 +177,8 @@ public class MemberAdminController {
         if (reportDashboardRepository.isDismissedInCommentAbuseReportDashboard(commentId)) {
             throw new ExistsValueException(EXISTS_COMMENT_ABUSE_REPORT_DISMISSED, "status");
         }
+        applicationEventPublisher.publishEvent(
+                CommentAbuseReportDismissEvent.create(postId.getValue(), commentPath.getValue()));
         return reportDashboardRepository.dismissCommentAbuseReport(commentId);
     }
 
@@ -123,7 +190,8 @@ public class MemberAdminController {
         if (reportDashboardRepository.isApprovedInCommentAbuseReportDashboard(commentId)) {
             throw new ExistsValueException(EXISTS_COMMENT_ABUSE_REPORT_BLINDED, "status");
         }
-        applicationEventPublisher.publishEvent(CommentAbuseReportApproveEvent.create(postId.getValue(), commentPath.getValue()));
+        applicationEventPublisher.publishEvent(
+                CommentAbuseReportApproveEvent.create(postId.getValue(), commentPath.getValue()));
         return reportDashboardRepository.approveCommentAbuseReport(commentId);
     }
 }
